run-vmtest: rewrite test list merging script
