+++ run-vmtest/tests/normalize_bpf_test_names/run-test.sh
@@ -5,6 +5,6 @@ set -euo pipefail
 GITHUB_ACTION_PATH=$(realpath ../..)
 
 rm -f output.txt
-cat input.txt | $GITHUB_ACTION_PATH/merge_test_lists.py 2>&1 > output.txt
+$GITHUB_ACTION_PATH/normalize_bpf_test_names.py input.txt 2>&1 > output.txt
 diff expected-output.txt output.txt
